client/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
colorlog==6.10.1
fastapi==0.136.0
h11==0.16.0
httptools==0.7.1
idna==3.11
pydantic==2.13.2
pydantic_core==2.46.2
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
